\name{Kei}
\alias{Kei}
\title{Wrapper function to return Kei}
\usage{
  Kei(...)
}
\arguments{
  \item{...}{arguments passed to \code{\link{Keir}}}
}
\description{
  Uses \code{\link{Keir}} and extracts imaginary component,
  i.e. 'Kei'
}
\examples{
Kei(1:10)
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\references{
  \url{http://mathworld.wolfram.com/Kei.html}
}
\seealso{
  \code{\link{Keir}}, \code{\link{Ker}}
}

