\name{Ker}
\alias{Ker}
\title{Wrapper function to return Ker}
\usage{
  Ker(...)
}
\arguments{
  \item{...}{arguments passed to \code{\link{Keir}}}
}
\description{
  Uses \code{\link{Keir}} and extracts real component, i.e.
  'Ker'
}
\examples{
Ker(1:10)
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\references{
  \url{http://mathworld.wolfram.com/Ker.html}
}
\seealso{
  \code{\link{Keir}}, \code{\link{Kei}}
}

