\name{Bei}
\alias{Bei}
\title{Wrapper function to return Bei}
\usage{
  Bei(...)
}
\description{
  Uses \code{\link{Beir}} and extracts imaginary component,
  i.e. 'Bei'
}
\examples{
Bei(1:10)
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\references{
  \link{http://mathworld.wolfram.com/Bei.html}
}
\seealso{
  \code{\link{Beir}}, \code{\link{Ber}}
}

