\name{Ber}
\alias{Ber}
\title{Wrapper function to return Ber}
\usage{
  Ber(...)
}
\description{
  Uses \code{\link{Beir}} and extracts real component, i.e.
  'Ber'
}
\examples{
Ber(1:10)
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\references{
  \link{http://mathworld.wolfram.com/Ber.html}
}
\seealso{
  \code{\link{Beir}}, \code{\link{Bei}}
}

