\name{Kei}
\alias{Kei}
\title{Wrapper function to return Kei}
\usage{
  Kei(...)
}
\description{
  Uses \code{\link{Kelvin}} and extracts imaginary
  component, i.e. 'Kei'
}
\examples{
Kei(1:10)
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\references{
  \link{http://mathworld.wolfram.com/Kei.html}
}
\seealso{
  \code{\link{Kelvin}}, \code{\link{Ker}}
}

