\name{Ker}
\alias{Ker}
\title{Wrapper function to return Ker}
\usage{
  Ker(...)
}
\description{
  Uses \code{\link{Kelvin}} and extracts real component,
  i.e. 'Ker'
}
\examples{
Ker(1:10)
}
\author{
  Andrew Barbour <andy.barbour@gmail.com>
}
\references{
  \link{http://mathworld.wolfram.com/Ker.html}
}
\seealso{
  \code{\link{Kelvin}}, \code{\link{Kei}}
}

